Got the special test cases passing

+++ src/addressingModes.c
@@ -4,8 +4,13 @@
 // The addressing modes do not offset the PC, this is handled by the CPU
 
 int ABS_A(int PC) { return (readByte(PC) + ((int)readByte(PC + 1) << 8) & 0xFFFF); }
-int ABS_INDEX_X(int PC) {return (ABS_A(PC) + readByte(getCPU_XRegister())) & 0xFFFF;}
-int ABS_INDEX_Y(int PC) {return (ABS_A(PC) + readByte(getCPU_YRegister())) & 0xFFFF;}
+int ABS_INDEX_X(int PC) {
+    //printf("ABS_PC")
+    return (ABS_A(PC) + (int)readByte(getCPU_XRegister())) & 0xFFFF;
+}
+int ABS_INDEX_Y(int PC) {
+    return (ABS_A(PC) + (int)readByte(getCPU_YRegister())) & 0xFFFF;
+}
 int ACC(int PC) {return getCPU_Accumulator();}
 int IMM(int PC) {return PC;}
 int IMP(int PC) {return -1;}
